indexer/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
pydantic==2.12.5
pydantic-settings==2.12.0

docling==2.70.0
docling-haystack==0.1.1

haystack-ai==2.21.0
ollama-haystack==5.3.0
pgvector-haystack==5.4.0
sentence-transformers==5.2.0